refactor(app): moves maps schema into shared schemas file
